Separated ValueWrapper into separate module.
